separate contexts into index.ts and types.ts files
