+++ frontend/src/Router.tsx
@@ -2,7 +2,7 @@ import React, { ReactElement } from "react";
 import { BrowserRouter, Switch, Route } from "react-router-dom";
 
 import NavigationBar from "./components/navigation/NavigationBar";
-import { PromotionsListProvider } from "./contexts/PromotionsList";
+import { PromotionsListProvider } from "./contexts/PromotionsListContext";
 import Home from "./screens/Home";
 import MyPromotions from "./screens/MyPromotions";
